git-svn-id: svn://scm.gforge.inria.fr/svn/openvibe@1812 c330d7e9-fc0c-0410-a5b3-fd85c6f5aa8f

+++ openvibe-plugins/signal-processing-gpl/trunc/src/box-algorithms/filter/ovpCBoxAlgorithmXDAWNSpatialFilterTrainer.cpp
@@ -402,7 +402,7 @@ boolean CBoxAlgorithmXDAWNSpatialFilterTrainer::process(void)
 				}
 			}
 			fprintf(l_pFile, "</SettingValue>\n");
-			fprintf(l_pFile, "\t<SettingValue>%d</SettingValue>\n", m_ui64FilterDimension);
+			fprintf(l_pFile, "\t<SettingValue>%d</SettingValue>\n", uint32(m_ui64FilterDimension));
 			fprintf(l_pFile, "\t<SettingValue>%d</SettingValue>\n", l_ui32ChannelCount);
 			fprintf(l_pFile, "</OpenViBE-SettingsOverride>\n");
 			fclose(l_pFile);
